Moved Node.JS tests to dev/tests/.
